feat(pipeline): implement pipe support
feat(tokenizer): implement tokenizer rules for pipes
feat(parser): introduce pipeline rules
feat(execute): implement execution of double pipeline
feat(main): re-enable redirection handling for single commands
feat(pipelines): implement pipe with builtins logic
refactor(pipeline): refactor parser and executor to support multi-command

+++ src/argvec.c
@@ -35,7 +35,7 @@ void append_arg_end(ArgVec* argv) {
 }
 
 void free_argvec(ArgVec* argv) {
-    for (int i = 0; i < argv->count; i++) {
+    for (size_t i = 0; i < argv->count; i++) {
         free(argv->args[i]);
     }
     free(argv->args);
