+++ src/redirect.c
@@ -1,12 +1,11 @@
 #include <stdlib.h>
 #include <string.h>
 
-#include "redirect.h"
 #include "error.h"
 #include "command.h"
 #include "tokenizer.h"
 
-void handle_redirect(Command* command, TokenArray* tokens, int* index, int* start, Token* redirect_token) {
+void handle_redirect(Command* command, TokenArray* tokens, size_t* index, size_t* start, Token* redirect_token) {
     // find path token
     int path_index = *index + 1;
     while(path_index < tokens->count && tokens->tokens[path_index].type == TOKEN_WHITESPACE) {
